move CloPos into core
