+++ src/main/java/io/github/opencubicchunks/cubicchunks/server/level/CubicServerLevel.java
@@ -1,7 +1,7 @@
 package io.github.opencubicchunks.cubicchunks.server.level;
 
 import io.github.opencubicchunks.cc_core.annotation.UsedFromASM;
-import io.github.opencubicchunks.cubicchunks.world.level.chunklike.CloPos;
+import io.github.opencubicchunks.cc_core.world.level.CloPos;
 
 public interface CubicServerLevel {
     @UsedFromASM
