+++ src/main/java/io/github/opencubicchunks/cubicchunks/server/level/CubicTaskPriorityQueueSorter.java
@@ -4,7 +4,7 @@ import java.util.function.IntConsumer;
 import java.util.function.IntSupplier;
 
 import io.github.opencubicchunks.cc_core.annotation.UsedFromASM;
-import io.github.opencubicchunks.cubicchunks.world.level.chunklike.CloPos;
+import io.github.opencubicchunks.cc_core.world.level.CloPos;
 
 public interface CubicTaskPriorityQueueSorter {
     @UsedFromASM
