+++ src/main/java/io/github/opencubicchunks/cubicchunks/server/level/progress/CubicChunkProgressListener.java
@@ -2,7 +2,7 @@ package io.github.opencubicchunks.cubicchunks.server.level.progress;
 
 import javax.annotation.Nullable;
 
-import io.github.opencubicchunks.cubicchunks.world.level.chunklike.CloPos;
+import io.github.opencubicchunks.cc_core.world.level.CloPos;
 import net.minecraft.world.level.chunk.ChunkStatus;
 
 public interface CubicChunkProgressListener {
