+++ src/test/java/io/github/opencubicchunks/cubicchunks/test/world/level/cube/TestLevelCube.java
@@ -15,9 +15,9 @@ import java.util.Map;
 import java.util.Random;
 import java.util.Set;
 
+import io.github.opencubicchunks.cc_core.world.level.CloPos;
 import io.github.opencubicchunks.cc_core.api.CubicConstants;
 import io.github.opencubicchunks.cubicchunks.testutils.BaseTest;
-import io.github.opencubicchunks.cubicchunks.world.level.chunklike.CloPos;
 import io.github.opencubicchunks.cubicchunks.world.level.cube.LevelCube;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.block.Blocks;
